add session and validation
