+++ app/src/main/java/com/mits/kakaroto/listmovieapp/main/MainActivity.java
@@ -6,16 +6,14 @@ import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
-import android.util.Log;
 import android.view.View;
-import android.widget.Toast;
 
-import com.mits.kakaroto.listmovieapp.model.AddMovieActivity;
-import com.mits.kakaroto.listmovieapp.model.DbHandlerTableMovies;
-import com.mits.kakaroto.listmovieapp.model.UpdateMovieActivity;
+import com.mits.kakaroto.listmovieapp.movie.AddMovieActivity;
+import com.mits.kakaroto.listmovieapp.database.DatabaseHandler;
+import com.mits.kakaroto.listmovieapp.movie.UpdateMovieActivity;
 import com.mits.kakaroto.listmovieapp.R;
 import com.mits.kakaroto.listmovieapp.adapter.MovieAdapter;
-import com.mits.kakaroto.listmovieapp.model.Movie;
+import com.mits.kakaroto.listmovieapp.movie.Movie;
 import com.mits.kakaroto.listmovieapp.utility.RecyclerTouchListener;
 import com.mits.kakaroto.listmovieapp.utility.SpacesItemDecoration;
 
@@ -28,24 +26,23 @@ public class MainActivity extends AppCompatActivity {
     private Intent intent;
     public static final int RESULT_ADD = 2;
     public static final int RESULT_UPDATE = 3;
-    private DbHandlerTableMovies db;
-    private List<Movie> list;
+    private DatabaseHandler tblMovie;
     private int pos;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        db = new DbHandlerTableMovies(this);
-
+        tblMovie = new DatabaseHandler(this);
         recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
+
         initRecyclerView();
 
     }
 
     public void initRecyclerView() {
-        list = db.getAllMovies();
-
+        List<Movie> list = tblMovie.getAllMovies();
         adapter = new MovieAdapter(list);
 
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
@@ -87,7 +84,7 @@ public class MainActivity extends AppCompatActivity {
         @Override
         public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction) {
             int position = viewHolder.getAdapterPosition();
-            db.deleteMovie(adapter.getItem(position));
+            tblMovie.deleteMovie(adapter.getItem(position));
             adapter.remove(position);
         }
     };
@@ -99,20 +96,18 @@ public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        String TAG = "";
         if (resultCode == RESULT_ADD) {
-            Log.d(TAG, "Success add movie");
             Movie movie = data.getParcelableExtra("data_add");
-            db.addMovies(movie);
+            tblMovie.addMovies(movie);
             adapter.insert(movie);
             recyclerView.scrollToPosition(0);
         } else if (resultCode == RESULT_UPDATE) {
             Movie movie = data.getParcelableExtra("data_update");
+            tblMovie.updateMovie(new Movie(movie.getId(), movie.getTitle(), movie.getGenre(),
+                    movie.getYear(), movie.getCountry(), movie.getDuration(),
+                    movie.getImageAddrees()));
 
-            db.updateMovie(new Movie(movie.getId(), movie.getTitle(), movie.getGenre(), movie.getYear(),
-                    movie.getCountry(), movie.getDuration(), movie.getImageAddrees()));
             adapter.update(pos, movie);
-            Log.d(TAG, "Update data Id : " + String.valueOf(movie.getId()));
             recyclerView.scrollToPosition(0);
         }
     }
