+++ app/src/main/java/com/mits/kakaroto/listmovieapp/session/DashboardActivity.java
@@ -7,20 +7,22 @@ import android.view.View;
 import android.widget.TextView;
 import com.mits.kakaroto.listmovieapp.R;
 import com.mits.kakaroto.listmovieapp.main.MainActivity;
+import com.mits.kakaroto.listmovieapp.database.DatabaseHandler;
+import com.mits.kakaroto.listmovieapp.user.User;
 
 import java.util.List;
 
 public class DashboardActivity extends AppCompatActivity {
     private SessionManager sessionManager;
 
-    private DbHandlerTableUsers tblUser;
+    private DatabaseHandler tblUser;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         sessionManager = SessionManager.getInstance();
         setContentView(R.layout.activity_dashbord);
-        tblUser = new DbHandlerTableUsers(this);
+        tblUser = new DatabaseHandler(this);
 
     }
 
@@ -30,15 +32,16 @@ public class DashboardActivity extends AppCompatActivity {
         String result = "";
         TextView tvUser = (TextView) findViewById(R.id.tv_user);
 
-        User user_ses = sessionManager.getUser();
+        String sessionEmail = sessionManager.getEmail();
+        String sesseionPass = sessionManager.getPass();
+
         List<User> userList = tblUser.getAllUser();
         for (User user : userList) {
-            if(user.getEmail().equals(user_ses.getEmail()) &&
-                    user.getPassword().equals(user_ses.getPassword())){
+            if(user.getEmail().equals(sessionEmail) &&
+                    user.getPassword().equals(sesseionPass)){
                 result = " Name : " + user.getName() + "\n Email : " + user.getEmail() + "\n Address : " +
                         user.getAddress() + "\n Phone : " + user.getPhone() + "\n" +
-                        " Gender : " + user.getGender() + "\n Password : " + user.getPassword() + "\n" +
-                        " \n \n \n User : "+user_ses.getEmail()+" pass : "+user_ses.getPassword();
+                        " Gender : " + user.getGender() + "\n Password : " + user.getPassword();
             }
         }
         tvUser.setText(result);
