+++ app/src/main/res/layout/activity_register_user.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.mits.kakaroto.listmovieapp.session.RegisterActivity">
+    tools:context="com.mits.kakaroto.listmovieapp.user.RegisterUserActivity">
