+++ app/src/main/res/layout/activity_update_movie.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:orientation="vertical"
-    tools:context="com.mits.kakaroto.listmovieapp.model.UpdateMovieActivity"
+    tools:context="com.mits.kakaroto.listmovieapp.movie.UpdateMovieActivity"
     >
 
     <EditText
@@ -17,6 +17,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:hint="@string/author"
+        android:inputType="textCapWords"
         android:layout_marginBottom="@dimen/activity_vertical_margin"
         />
 
@@ -25,6 +26,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:hint="@string/genre"
+        android:inputType="textCapWords"
         android:layout_marginBottom="@dimen/activity_vertical_margin"
         />
 
@@ -33,6 +35,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:hint="@string/year"
+        android:inputType="number"
         android:layout_marginBottom="@dimen/activity_vertical_margin"
         />
 
@@ -41,6 +44,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:hint="@string/country"
+        android:inputType="textCapWords"
         android:layout_marginBottom="@dimen/activity_vertical_margin"
         />
 
@@ -50,6 +54,7 @@
         android:layout_height="wrap_content"
         android:layout_marginBottom="@dimen/activity_vertical_margin"
         android:hint="@string/duration"
+        android:inputType="time"
         />
