+++ app/src/main/res/layout/activity_update_user.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:orientation="vertical"
-    tools:context="com.mits.kakaroto.listmovieapp.session.UpdateActivity">
+    tools:context="com.mits.kakaroto.listmovieapp.user.UpdateUserActivity">
 
     <EditText
         android:id="@+id/et_updateName"
